Move abstract class to namespace.
